added api folder and moved the api logic to one file
